soperator: add hacky slurm-jobs dashboard from accounting DB
I had no time to finish and test it properly, but nevertheless.

Remember: mariadb-password is not automatically copied from soperator namespace
to monitoring-system, you have to copy it yourself and then restart
grafana instance in order for it to grab secret into its env.

kyverno or some other solution could be implemented in the future to
automate this part.